backend/requirements.txt
# 火源文案智能体 - Python Dependencies
# Install: pip install -r requirements.txt

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP Client
httpx>=0.26.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0

# Optional: Development Tools
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# black>=24.1.0
# ruff>=0.1.0


